OG commit

+++ libft.h/ft_strlcpy.c
@@ -6,15 +6,15 @@
 /*   By: lduheron <lduheron@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 09:22:55 by lduheron          #+#    #+#             */
-/*   Updated: 2022/11/11 09:23:00 by lduheron         ###   ########.fr       */
+/*   Updated: 2022/11/28 20:08:04 by lduheron         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlcpy(char *dst, const char *src, unsigned int size)
+size_t	ft_strlcpy(char *dst, const char *src, unsigned int size)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (size > 0)
